Fixed previous issues.
This was due to the use of the mulhi and mul24 instructions.
They can probably be used correctly but for now we will just get by with
a normal add and multiply.

This was causing issues in other filters as well, such as LUT,
     MaskedMerge, and Merge.

Now, luckily, all filters work, including Expr.

The use of the mul* instructions will need to be investigated in the
future.