Fixed an issue with the constant ops copy.
This was a result of the fact that multiple Expr filter calls can be
made, very close together. Since the original ops copy was in the
create() section, ops would be over written as they were being used.

Now, we force the copy upon each getFrame(), which is much more stable.
This work needs to be improved a bit, to ensure that we copy the ops as
few times as possible. Maybe limiting it to the init() phase will help,
or using stream synchronization to prevent over writing.

That is another caveat of the constant memory approach. You cannot
safely conduct multiple Expr filters simultaneously on the GPU. They
must be seperated to some extent. The current architecture pretty much
ensures this, but does not guarantee it...exactly.

More testing is required.