Optimized constant memory usage in Expr.
This was done to reduce the constant memory footprint of Expr,
as before it was wasting a large amount of space for filter calls
that only filter one or two planes.

This is currently UNTESTED, but should be pretty close to stable.
It has the side benefit of reducing the code size and cleaning
some redundant portions.

Signed-off-by: Austin Dworaczyk Wiltshire <adw.wiltshire@gmail.com>