Changed all stream retrievel to retrieve from the destination frame.
For some interesting reason, pulling from a source frame can result in
corruption when filtering. It's likely due to a race condition when
dealing with multiple input clips and memory copies to the device.

Further testing is required to make sure that all corruption issues are
mitigated by pulling the stream selection from the destination frame,
vs. using explicit synchronization.