Switched Expr back over to using global memory for ops.
Theoretically on Kepler, the ops should be loaded into the readonly
cache. Not quite clear on how to detect this though.

This makes Expr a lot more adaptable and elimnates a reliance on
compile time limitations for operations.

Does constant memory make Expr faster? Yes, very much so.
Is it worth the trade of in usability? I'm not so sure.

Plus, constant memory use is only a ~5% performance boost.
We can always switch it back, but I will take stability over
raw performance any day.